requirements.txt
imageio-ffmpeg>=0.4.0
inflect>=0.3.0
macos-tags>=1.5.0
patool>=1.15.0
pillow>=10.4.0
titlecase>=2.4.0
Wand>=0.6.0